+++ tests/Unit/ArrayDefinitionTest.php
@@ -503,7 +503,7 @@ final class ArrayDefinitionTest extends TestCase
 
     public function testMagicMethods(): void
     {
-        $definiton = ArrayDefinition::fromConfig([
+        $definition = ArrayDefinition::fromConfig([
             'class' => Recorder::class,
             'first()' => [],
             '$second' => null,
@@ -511,7 +511,7 @@ final class ArrayDefinitionTest extends TestCase
             '$fourth' => 'hello',
         ]);
 
-        $object = $definiton->resolve(new SimpleContainer());
+        $object = $definition->resolve(new SimpleContainer());
 
         $this->assertSame(
             [
@@ -521,6 +521,27 @@ final class ArrayDefinitionTest extends TestCase
                 'Set $fourth to string',
             ],
             $object->getEvents()
+        );
+    }
+
+    public function testMultipleCall()
+    {
+        $definition = ArrayDefinition::fromConfig([
+            'class' => Recorder::class,
+            'test()' => [1],
+            'test()_2' => [2],
+            'test()_3' => [3],
+        ]);
+
+        $object = $definition->resolve(new SimpleContainer());
+
+        $this->assertSame(
+            [
+                'Call test(int)',
+                'Call test(int)',
+                'Call test(int)',
+            ],
+            $object->getEvents(),
         );
     }
 }