+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.2.1 under development
 
 - Bug #86: Fix crash when intersection types are used (@vjik)
+- New #90: Allow multiple method call in array definition (@vjik)
 
 ## 3.2.0 February 12, 2023